+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [vNext]
+### Updated
+- Added more cached parameters into Scroll providers. [Trello 1854](https://trello.com/c/Sdzfuue2)
+
 ## [3.197.0]
 ### Added
 - A new api for padding coded regions. [Trello 2507](https://trello.com/c/LhysNDqu)